renaming of adapters
